Small refactoring of visitor interfaces
