+++ com.langtoun.oastypes/src/main/java/com/langtoun/oastypes/visitor/OASTypeDispatcher.java
@@ -1,4 +1,4 @@
-package com.langtoun.oastypes.util;
+package com.langtoun.oastypes.visitor;
 
 import com.langtoun.oastypes.OASType;
